requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.4
django-rq==3.0.0
django-timescaledb==0.2.13
djangorestframework==3.15.2
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rq==2.0.0
six==1.17.0
sqlparse==0.5.3
supervisor==4.2.5
typing_extensions==4.12.2
urllib3==2.2.3
